Productos completo, con un poco mas de logica real, ya se puede trabajar mejor el modulo de ventas

+++ src/main/java/com/pillone/pillone/repository/ProductosRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductosRepository extends JpaRepository<Productos, Integer> {
+public interface ProductosRepository extends JpaRepository<Productos, Long> {
 }
